app/requirements.txt
Django==3.2.6
gunicorn==20.1.0
psycopg2-binary==2.9.1
beautifulsoup4==4.11.1
requests==2.28.0
mysqlclient==2.1.1
django_cron==0.5.1
django-tagulous==1.3.3
django-crontab==0.7.1